src/requirements.txt
django
apprise
gevent
gunicorn
django-cors-headers
python-socketio
django-session-timeout
mozilla-django-oidc
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-filter
drf-yasg
celery
django-celery-beat
django-celery-results
requests
psycopg2-binary
django-db-geventpool
sentry-sdk
dj-rest-auth
python-json-logger
django-storages
boto3
splunk_handler
zeep
django-prometheus
prometheus-client